backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pymupdf
docx2txt
google-generativeai
scikit-learn
nltk
requests
python-dotenv
